Support for absolute value mode has been improved.
git-svn-id: svn+ssh://cv077/home/ueshiba/svnroot/trunk/src@2031 3ef1b8b6-fb57-41bf-88d9-9a5bd220bf38

+++ testcam3/MyCmdWindow.h
@@ -235,7 +235,7 @@ MyCmdWindow<PIXEL>::initializeMovie()
 	    _canvases[i]->resize();
     }
 
-    int	props[] = {0, _movie.nframes() - 1, 1};
+    int	props[] = {0, static_cast<int>(_movie.nframes() - 1), 1};
     _captureCmd.setProp(c_StatusMovie, props);
 
     repaintCanvases();
